+++ src/heliax/__init__.py
@@ -151,7 +151,7 @@ from .training import Trainer
 from .workspace import Workspace
 from .workspace import workspace as workspace_context
 
-__version__ = "0.6.16"
+__version__ = "0.6.17"
 
 __all__ = [
     "GELU",